Partial geo database sample
The geo_database sample_50.pkl file contains a random sample of the final data expected to be database in the project. The other modifications correspond to the union between the information in the IGAC dictionary and that found in the 7 layers defined for Chaparral in the EDA developed by Sergio.